fix: only upload if action is demo

+++ .travis/after_success_coverage.sh
@@ -1,24 +1,28 @@
 #!/usr/bin/env bash
 
-echo "COVERAGE"
-cd "$TRAVIS_BUILD_DIR"
+if [ "$ACTION" == 'demo' ]; then
+    echo "COVERAGE"
+    cd "$TRAVIS_BUILD_DIR"
 
-rm -rf .static/cmf/coverage
-cp -R packages/cmf/coverage/lcov-report/ .static/cmf/coverage
-echo "✓ Copy cmf coverage to .static"
+    rm -rf .static/cmf/coverage
+    cp -R packages/cmf/coverage/lcov-report/ .static/cmf/coverage
+    echo "✓ Copy cmf coverage to .static"
 
-rm -rf .static/components/coverage
-cp -R packages/components/coverage/lcov-report/ .static/components/coverage
-echo "✓ Copy components coverage to .static"
+    rm -rf .static/components/coverage
+    cp -R packages/components/coverage/lcov-report/ .static/components/coverage
+    echo "✓ Copy components coverage to .static"
 
-rm -rf .static/containers/coverage
-cp -R packages/containers/coverage/lcov-report/ .static/containers/coverage
-echo "✓ Copy containers coverage to .static"
+    rm -rf .static/containers/coverage
+    cp -R packages/containers/coverage/lcov-report/ .static/containers/coverage
+    echo "✓ Copy containers coverage to .static"
 
-rm -rf .static/forms/coverage
-cp -R packages/forms/coverage/lcov-report/ .static/forms/coverage
-echo "✓ Copy forms coverage to .static"
+    rm -rf .static/forms/coverage
+    cp -R packages/forms/coverage/lcov-report/ .static/forms/coverage
+    echo "✓ Copy forms coverage to .static"
 
-rm -rf .static/logging/coverage
-cp -R packages/logging/coverage/lcov-report/ .static/logging/coverage
-echo "✓ Copy logging coverage to .static"
+    rm -rf .static/logging/coverage
+    cp -R packages/logging/coverage/lcov-report/ .static/logging/coverage
+    echo "✓ Copy logging coverage to .static"
+else
+    echo "✓ no COVERAGE to copy"
+fi
